requirements.txt
# Core dependencies
aiogram>=2.25.1,<3.0.0
python-dotenv>=1.0.0
aiohttp>=3.8.5

# Database
aiosqlite>=0.19.0

# Optional payment providers (uncomment when needed):
# stripe>=5.0.0
# paypalrestsdk>=1.13.1
